requirements.txt
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
